% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{parse_html}
\alias{parse_html}
\title{Convert arbitrary HTML to htmltools tags}
\usage{
parse_html(html)
}
\arguments{
\item{html}{\code{chr} HTML or path to html file}
}
\value{
\code{shiny.tag.list}
}
\description{
Convert arbitrary HTML to htmltools tags
}
\examples{
parse_html('<div id = "test"></div>')
}
